feature/ATC-181 - Creates AircraftCommander class to run commands on an AircraftInstance * Removes commands section in AircraftInstanceModel * Adds AircraftCommander instance to App that gets passed into * InputController * Removes no-undef rule from .eslintrc
