+++ .eslintrc
@@ -31,7 +31,6 @@
         "no-restricted-syntax": 0,
         "no-shadow": 0,
         "no-mixed-operators": 0,
-        "no-undef": 0,
         "no-underscore-dangle": 0,
         "object-shorthand": 0
     }
