+++ src/assets/scripts/client/aircraft/AircraftInstanceModel.js
@@ -4,15 +4,12 @@ import _forEach from 'lodash/forEach';
 import _get from 'lodash/get';
 import _has from 'lodash/has';
 import _isEqual from 'lodash/isEqual';
-import _isNaN from 'lodash/isNaN';
 import _isNil from 'lodash/isNil';
 import _isString from 'lodash/isString';
 import _map from 'lodash/map';
 import _without from 'lodash/without';
 import AircraftFlightManagementSystem from './FlightManagementSystem/AircraftFlightManagementSystem';
 import AircraftStripView from './AircraftStripView';
-import Waypoint from './FlightManagementSystem/Waypoint';
-import RouteModel from '../airport/Route/RouteModel';
 import { speech_say } from '../speech';
 import { tau, radians_normalize, angle_offset } from '../math/circle';
 import { round, abs, sin, cos, extrapolate_range_clamp, clamp } from '../math/core';
@@ -31,60 +28,20 @@ import {
     point_in_area
 } from '../math/vector';
 import {
-    radio_cardinalDir_names,
     digits_decimal,
     groupNumbers,
     radio_runway,
-    radio_heading,
     radio_spellOut,
-    radio_altitude,
-    radio_trend,
-    getCardinalDirection
+    radio_altitude
 } from '../utilities/radioUtilities';
 import { km, radiansToDegrees, degreesToRadians, heading_to_string } from '../utilities/unitConverters';
 import {
     FLIGHT_MODES,
     FLIGHT_CATEGORY,
-    WAYPOINT_NAV_MODE,
-    FP_LEG_TYPE
+    WAYPOINT_NAV_MODE
 } from '../constants/aircraftConstants';
 import { SELECTORS } from '../constants/selectors';
 import { GAME_EVENTS } from '../game/GameController';
-
-/**
- * Enum of commands and thier corresponding function.
- *
- * Used to build a call to the correct function when a UI command, or commands,
- * for an aircraft have been issued.
- *
- * @property COMMANDS
- * @type {Object}
- * @final
- */
-const COMMANDS = {
-    abort: 'runAbort',
-    altitude: 'runAltitude',
-    clearedAsFiled: 'runClearedAsFiled',
-    climbViaSID: 'runClimbViaSID',
-    debug: 'runDebug',
-    delete: 'runDelete',
-    descendViaSTAR: 'runDescendViaSTAR',
-    direct: 'runDirect',
-    fix: 'runFix',
-    flyPresentHeading: 'runFlyPresentHeading',
-    heading: 'runHeading',
-    hold: 'runHold',
-    land: 'runLanding',
-    moveDataBlock: 'runMoveDataBlock',
-    route: 'runRoute',
-    reroute: 'runReroute',
-    sayRoute: 'runSayRoute',
-    sid: 'runSID',
-    speed: 'runSpeed',
-    star: 'runSTAR',
-    takeoff: 'runTakeoff',
-    taxi: 'runTaxi'
-};
 
 /**
  * @property FLIGHT_RULES
@@ -603,1012 +560,6 @@ export default class Aircraft {
         this.$html.hide(600);
     }
 
-
-    // TODO: move aircraftCommands to a new class
-    /**
-     * @for AircraftInstanceModel
-     * @method runCommands
-     * @param commands
-     */
-    runCommands(commands) {
-        if (!this.inside_ctr) {
-            return true;
-        }
-
-        let response = [];
-        let response_end = '';
-        const deferred = [];
-
-        for (let i = 0; i < commands.length; i++) {
-            const command = commands[i][0];
-            const args = commands[i].splice(1);
-
-            if (command === FLIGHT_MODES.TAKEOFF) {
-                deferred.push([command, args]);
-                continue;
-            }
-
-            let retval = this.run(command, args);
-
-            if (retval) {
-                if (!_has(retval[1], 'log') || !_has(retval[1], 'say')) {
-                    // TODO: reassigning a value using itself is dangerous. this should be re-wroked
-                    retval = [
-                        retval[0],
-                        {
-                            log: retval[1],
-                            say: retval[1]
-                        }
-                    ];
-                }
-
-                response.push(retval[1]);
-
-                if (retval[2]) {
-                    response_end = retval[2];
-                }
-            }
-        }
-
-        for (let i = 0; i < deferred.length; i += 1) {
-            const command = deferred[i][0];
-            const args = deferred[i][1];
-            const retval  = this.run(command, args);
-
-            if (retval) {
-                // TODO: fix the logic here this very purposly using `!=`. length is not an object and thus,
-                // never null but by using coercion it evaluates to falsey if its not an array
-                // true if array, and not log/say object
-                if (retval[1].length != null) {
-                    // make into log/say object
-                    retval[1] = {
-                        say: retval[1],
-                        log: retval[1]
-                    };
-                }
-
-                response.push(retval[1]);
-            }
-        }
-
-        if (commands.length === 0) {
-            response = [{
-                say: 'not understood',
-                log: 'not understood'
-            }];
-            response_end = 'say again';
-        }
-
-        if (response.length >= 1) {
-            if (response_end) {
-                response_end = `, ${response_end}`;
-            }
-
-            const r_log = _map(response, (r) => r.log).join(', ');
-            const r_say = _map(response, (r) => r.say).join(', ');
-
-            window.uiController.ui_log(`${this.getCallsign()}, ${r_log} ${response_end}`);
-            speech_say([
-                { type: 'callsign', content: this },
-                { type: 'text', content: `${r_say} ${response_end}` }
-            ]);
-        }
-
-        this.updateStrip();
-
-        return true;
-    }
-
-    /**
-     * @for AircraftInstanceModel
-     * @method run
-     * @param command
-     * @param data
-     * @return {function}
-     */
-    run(command, data) {
-        let call_func;
-
-        if (COMMANDS[command]) {
-            call_func = COMMANDS[command];
-        }
-
-        if (!call_func) {
-            return ['fail', 'not understood'];
-        }
-
-        return this[call_func](data);
-    }
-
-    /**
-     * @for AircraftInstanceModel
-     * @method runHeading
-     * @param data
-     */
-    runHeading(data) {
-        const airport = window.airportController.airport_get();
-        const direction = data[0];
-        let heading = data[1];
-        const incremental = data[2];
-        let amount = 0;
-        let instruction;
-
-        if (_isNaN(heading)) {
-            return ['fail', 'heading not understood'];
-        }
-
-        if (incremental) {
-            amount = heading;
-
-            if (direction === 'left') {
-                heading = radiansToDegrees(this.heading) - amount;
-            } else if (direction === 'right') {
-                heading = radiansToDegrees(this.heading) + amount;
-            }
-        }
-
-        // TODO: this probably shouldn't be the AircraftInstanceModel's job. this logic should belong somewhere else.
-        // Update the FMS
-        let wp = this.fms.currentWaypoint;
-        const leg = this.fms.currentLeg;
-        const f = this.fms.following;
-
-        if (wp.navmode === WAYPOINT_NAV_MODE.RWY) {
-            this.cancelLanding();
-        }
-
-        // already being vectored or holding. Will now just change the assigned heading.
-        if (wp.navmode === WAYPOINT_NAV_MODE.HEADING) {
-            this.fms.setCurrent({
-                altitude: wp.altitude,
-                navmode: WAYPOINT_NAV_MODE.HEADING,
-                heading: degreesToRadians(heading),
-                speed: wp.speed,
-                turn: direction,
-                hold: false
-            });
-        } else if (wp.navmode === WAYPOINT_NAV_MODE.HOLD) {
-            // in hold. Should leave the hold, and add leg for vectors
-            const index = this.fms.current[0] + 1;
-            const waypointToAdd = new Waypoint(
-                {
-                    altitude: wp.altitude,
-                    navmode: WAYPOINT_NAV_MODE.HEADING,
-                    heading: degreesToRadians(heading),
-                    speed: wp.speed,
-                    turn: direction,
-                    hold: false
-                },
-                airport
-            );
-
-            // add new Leg after hold leg
-            this.fms.insertLeg({
-                firstIndex: index,
-                waypoints: [waypointToAdd]
-            });
-
-            // move from hold leg to vector leg.
-            this.fms.nextWaypoint();
-        } else if (f.sid || f.star || f.awy) {
-            const waypointToAdd = new Waypoint(
-                {
-                    altitude: wp.altitude,
-                    navmode: WAYPOINT_NAV_MODE.HEADING,
-                    heading: degreesToRadians(heading),
-                    speed: wp.speed,
-                    turn: direction,
-                    hold: false
-                },
-                airport
-            );
-
-            // TODO: this should be an FMS class method that accepts a new `waypointToAdd`
-            // insert wp with heading at current position within the already active leg
-            leg.waypoints.splice(this.fms.current[1], 0, waypointToAdd);
-        } else if (leg.route !== '[radar vectors]') {
-            // needs new leg added
-            if (this.fms.atLastWaypoint()) {
-                const waypointToAdd = new Waypoint(
-                    {
-                        altitude: wp.altitude,
-                        navmode: WAYPOINT_NAV_MODE.HEADING,
-                        heading: degreesToRadians(heading),
-                        speed: wp.speed,
-                        turn: direction,
-                        hold: false
-                    },
-                    airport
-                );
-
-                this.fms.appendLeg({
-                    waypoints: [waypointToAdd]
-                });
-
-                this.fms.nextLeg();
-            } else {
-                const waypointToAdd = new Waypoint(
-                    {
-                        altitude: wp.altitude,
-                        navmode: WAYPOINT_NAV_MODE.HEADING,
-                        heading: degreesToRadians(heading),
-                        speed: wp.speed,
-                        turn: direction,
-                        hold: false
-                    },
-                    airport
-                );
-
-                this.fms.insertLegHere({
-                    waypoints: [waypointToAdd]
-                });
-            }
-        }
-
-        wp = this.fms.currentWaypoint;  // update 'wp'
-
-        // Construct the readback
-        instruction = 'fly heading';
-        if (direction) {
-            instruction = `turn ${direction} heading`;
-        }
-
-        const readback = {};
-        readback.log = `${instruction} ${heading_to_string(wp.heading)}`;
-        readback.say = `${instruction} ${radio_heading(heading_to_string(wp.heading))}`;
-
-        if (incremental) {
-            readback.log = `turn ${amount} degrees ${direction}`;
-            readback.say = `turn ${groupNumbers(amount)} degrees ${direction}`;
-        }
-
-        return ['ok', readback];
-    }
-
-    /**
-     * @for AircraftInstanceModel
-     * @method runAltitude
-     * @param data
-     */
-    runAltitude(data) {
-        const altitude = data[0];
-        let expedite = data[1];
-        const airport = window.airportController.airport_get();
-        const radioTrendAltitude = radio_trend('altitude', this.altitude, this.fms.altitudeForCurrentWaypoint());
-        const currentWaypointRadioAltitude = radio_altitude(this.fms.altitudeForCurrentWaypoint());
-
-        if ((altitude == null) || isNaN(altitude)) {
-            // FIXME: move this to it's own command. if expedite can be passed as a sole command it should be its own command
-            if (expedite) {
-                this.fms.setCurrent({ expedite: true });
-
-                return ['ok', `${radioTrendAltitude} ${this.fms.altitudeForCurrentWaypoint()} expedite`];
-            }
-
-            return ['fail', 'altitude not understood'];
-        }
-
-        if (this.mode === FLIGHT_MODES.LANDING) {
-            this.cancelLanding();
-        }
-
-        let ceiling = airport.ctr_ceiling;
-        if (window.gameController.game.option.get('softCeiling') === 'yes') {
-            ceiling += 1000;
-        }
-
-        this.fms.setAll({
-            // TODO: enumerate the magic numbers
-            altitude: clamp(round(airport.elevation / 100) * 100 + 1000, altitude, ceiling),
-            expedite: expedite
-        });
-
-        let isExpeditingString = '';
-        if (expedite) {
-            isExpeditingString = 'and expedite';
-        }
-
-        const readback = {
-            log: `${radioTrendAltitude} ${this.fms.altitudeForCurrentWaypoint()} ${isExpeditingString}`,
-            say: `${radioTrendAltitude} ${currentWaypointRadioAltitude} ${isExpeditingString}`
-        };
-
-        return ['ok', readback];
-    }
-
-    /**
-     * @for AircraftInstanceModel
-     * @method runClearedAsFiled
-     * @return {array}
-     */
-    runClearedAsFiled() {
-        if (!this.runSID([this.destination])) {
-            return [true, 'unable to clear as filed'];
-        }
-
-        const airport = window.airportController.airport_get();
-        const { name: procedureName } = airport.sidCollection.findRouteByIcao(this.destination);
-        const readback = {};
-
-        readback.log = `cleared to destination via the ${this.destination} departure, then as filed. Climb and ` +
-            `maintain ${airport.initial_alt}, expect ${this.fms.fp.altitude} 10 minutes after departure `;
-        readback.say = `cleared to destination via the ${procedureName} ` +
-            `departure, then as filed. Climb and maintain ${radio_altitude(airport.initial_alt)}, ` +
-            `expect ${radio_altitude(this.fms.fp.altitude)}, ${radio_spellOut('10')} minutes after departure'`;
-
-        return ['ok', readback];
-    }
-
-    /**
-     * @for AircraftInstanceModel
-     * @method runClimbViaSID
-     */
-    runClimbViaSID() {
-        if (this.fms.currentLeg.type !== FP_LEG_TYPE.SID || !this.fms.climbViaSID()) {
-            const isWarning = true;
-
-            window.uiController.ui_log(`${this.getCallsign()} unable to climb via SID`, isWarning);
-
-            return;
-        }
-
-        const airport = window.airportController.airport_get();
-        const { name: procedureName } = airport.sidCollection.findRouteByIcao(this.fms.currentLeg.route.procedure);
-        const readback = {
-            log: `climb via the ${this.fms.currentLeg.route.procedure} departure`,
-            say: `climb via the ${procedureName} departure`
-        };
-
-        return ['ok', readback];
-    }
-
-    /**
-     * @for AircraftInstanceModel
-     * @method runDescendViaSTAR
-     * @param data
-     * @return {boolean|undefined}
-     */
-    runDescendViaSTAR() {
-        if (!this.fms.descendViaSTAR() || !this.fms.following.star) {
-            const isWarning = true;
-            window.uiController.ui_log(`${this.getCallsign()}, unable to descend via STAR`, isWarning);
-
-            return;
-        }
-
-        const airport = window.airportController.airport_get();
-        const { name: procedureName } = airport.starCollection.findRouteByIcao(this.fms.currentLeg.route.procedure);
-        const readback = {
-            log: `descend via the ${this.fms.following.star} arrival`,
-            say: `descend via the ${procedureName} arrival`
-        };
-
-        return ['ok', readback];
-    }
-
-    /**
-     * @for AircraftInstanceModel
-     * @method runSpeed
-     * @param data
-     */
-    runSpeed(data) {
-        const speed = data[0];
-
-        if (_isNaN(speed)) {
-            return ['fail', 'speed not understood'];
-        }
-
-        const clampedSpeed = clamp(this.model.speed.min, speed, this.model.speed.max);
-        this.fms.setAll({ speed: clampedSpeed });
-
-        const radioTrendSpeed = radio_trend('speed', this.speed, this.fms.currentWaypoint.speed);
-        const readback = {
-            log: `${radioTrendSpeed} ${this.fms.currentWaypoint.speed}`,
-            say: `${radioTrendSpeed} ${radio_spellOut(this.fms.currentWaypoint.speed)}`
-        };
-
-        return ['ok', readback];
-    }
-
-    /**
-     * @for AircraftInstanceModel
-     * @method runHold
-     * @param data
-     */
-    runHold(data) {
-        const airport = window.airportController.airport_get();
-        let dirTurns = data[0];
-        let legLength = data[1];
-        let holdFix = data[2];
-        let holdFixLocation = null;
-        let inboundHdg;
-        // let inboundDir;
-
-        // TODO: this might be better handled from within the parser
-        if (dirTurns == null) {
-            // standard for holding patterns is right-turns
-            dirTurns = 'right';
-        }
-
-        // TODO: this might be better handled from within the parser
-        if (legLength == null) {
-            legLength = '1min';
-        }
-
-        // TODO: simplify this nested if.
-        if (holdFix !== null) {
-            holdFix = holdFix.toUpperCase();
-            holdFixLocation = airport.getFixPosition(holdFix);
-
-            if (!holdFixLocation) {
-                return ['fail', `unable to find fix ${holdFix}`];
-            }
-        }
-
-        if (this.isTakeoff() && !holdFix) {
-            return ['fail', 'where do you want us to hold?'];
-        }
-
-        // Determine whether or not to enter the hold from present position
-        if (holdFix) {
-            // holding over a specific fix (currently only able to do so on inbound course)
-            inboundHdg = vradial(vsub(this.position, holdFixLocation));
-
-            if (holdFix !== this.fms.currentWaypoint.fix) {
-                // not yet headed to the hold fix
-                this.fms.insertLegHere({
-                    type: 'fix',
-                    route: '[GPS/RNAV]',
-                    waypoints: [
-                        // proceed direct to holding fix
-                        new Waypoint(
-                            {
-                                fix: holdFix,
-                                altitude: this.fms.altitudeForCurrentWaypoint(),
-                                speed: this.fms.currentWaypoint.speed
-                            },
-                            airport
-                        ),
-                        // then enter the hold
-                        new Waypoint(
-                            {
-                                navmode: WAYPOINT_NAV_MODE.HOLD,
-                                speed: this.fms.currentWaypoint.speed,
-                                altitude: this.fms.altitudeForCurrentWaypoint(),
-                                fix: null,
-                                hold: {
-                                    fixName: holdFix,
-                                    fixPos: holdFixLocation,
-                                    dirTurns: dirTurns,
-                                    legLength: legLength,
-                                    inboundHdg: inboundHdg,
-                                    timer: null
-                                }
-                            },
-                            airport
-                        )
-                    ]
-                });
-            } else {
-                // TODO: this should be a `Waypoint`
-                // already currently going to the hold fix
-                // Force the initial turn to outbound heading when entering the hold
-                this.fms.appendWaypoint({
-                    navmode: WAYPOINT_NAV_MODE.HOLD,
-                    speed: this.fms.currentWaypoint.speed,
-                    altitude: this.fms.altitudeForCurrentWaypoint(),
-                    fix: null,
-                    hold: {
-                        fixName: holdFix,
-                        fixPos: holdFixLocation,
-                        dirTurns: dirTurns,
-                        legLength: legLength,
-                        inboundHdg: inboundHdg,
-                        timer: null
-                    }
-                });
-            }
-        } else {
-            // holding over present position (currently only able to do so on present course)
-            holdFixLocation = this.position; // make a/c hold over their present position
-            inboundHdg = this.heading;
-
-            // TODO: these aren't `Waypoints` and they should be
-            this.fms.insertLegHere({
-                type: 'fix',
-                waypoints: [
-                    { // document the present position as the 'fix' we're holding over
-                        navmode: WAYPOINT_NAV_MODE.FIX,
-                        fix: '[custom]',
-                        location: holdFixLocation,
-                        altitude: this.fms.altitudeForCurrentWaypoint(),
-                        speed: this.fms.currentWaypoint.speed
-                    },
-                    { // Force the initial turn to outbound heading when entering the hold
-                        navmode: WAYPOINT_NAV_MODE.HOLD,
-                        speed: this.fms.currentWaypoint.speed,
-                        altitude: this.fms.altitudeForCurrentWaypoint(),
-                        fix: null,
-                        hold: {
-                            fixName: holdFix,
-                            fixPos: holdFixLocation,
-                            dirTurns: dirTurns,
-                            legLength: legLength,
-                            inboundHdg: inboundHdg,
-                            timer: null
-                        }
-                    }
-                ]
-            });
-        }
-
-        // TODO: abstract to method `.getInboundCardinalDirection()`
-        const inboundDir = radio_cardinalDir_names[getCardinalDirection(radians_normalize(inboundHdg + Math.PI)).toLowerCase()];
-
-        if (holdFix) {
-            return ['ok', `proceed direct ${holdFix} and hold inbound, ${dirTurns} turns, ${legLength} legs`];
-        }
-
-        return ['ok', `hold ${inboundDir} of present position, ${dirTurns} turns, ${legLength} legs`];
-    }
-
-    /**
-     * @for AircraftInstanceModel
-     * @method runDirect
-     * @param data
-     */
-    runDirect(data) {
-        const fixname = data[0].toUpperCase();
-        // TODO replace with FixCollection
-        const fix = window.airportController.airport_get().getFixPosition(fixname);
-
-        if (!fix) {
-            return ['fail', `unable to find fix called ${fixname}`];
-        }
-
-        // remove intermediate fixes
-        if (this.mode === FLIGHT_MODES.TAKEOFF) {
-            this.fms.skipToFix(fixname);
-        } else if (!this.fms.skipToFix(fixname)) {
-            return ['fail', `${fixname} is not in our flightplan`];
-        }
-
-        return ['ok', `proceed direct ${fixname}`];
-    }
-
-    runFix(data) {
-        let last_fix;
-        let fail;
-        const fixes = _map(data, (fixname) => {
-            // TODO: this may beed to be the FixCollection
-            const fix = window.airportController.airport_get().getFixPosition(fixname);
-
-            if (!fix) {
-                fail = ['fail', `unable to find fix called ${fixname}`];
-
-                return;
-            }
-
-            // to avoid repetition, compare name with the previous fix
-            if (fixname === last_fix) {
-                return;
-            }
-
-            last_fix = fixname;
-
-            return fixname;
-        });
-
-        if (fail) {
-            return fail;
-        }
-
-        for (let i = 0; i < fixes.length; i++) {
-            // FIXME: use enumerated constant for type
-            this.fms.insertLegHere({ type: 'fix', route: fixes[i] });
-        }
-
-        if (this.mode !== FLIGHT_MODES.WAITING &&
-            this.mode !== FLIGHT_MODES.TAKEOFF &&
-            this.mode !== FLIGHT_MODES.APRON &&
-            this.mode !== FLIGHT_MODES.TAXI
-        ) {
-            this.cancelLanding();
-        }
-
-        return ['ok', `proceed direct ${fixes.join(', ')}`];
-    }
-
-    /**
-     * @for AircraftInstanceModel
-     * @method runFlyPresentHeading
-     * @param data
-     */
-    runFlyPresentHeading(data) {
-        this.cancelFix();
-        this.runHeading([null, radiansToDegrees(this.heading)]);
-
-        return ['ok', 'fly present heading'];
-    }
-
-    /**
-     * @for AircraftInstanceModel
-     * @method runSayRoute
-     * @param data
-     */
-    runSayRoute(data) {
-        return ['ok', {
-            log: `route: ${this.fms.fp.route.join(' ')}`,
-            say: 'here\'s our route'
-        }];
-    }
-
-    /**
-     * @for AircraftInstanceModel
-     * @method runSID
-     */
-    runSID(data) {
-        const airport = window.airportController.airport_get();
-        const { sidCollection } = airport;
-        const sidId = data[0];
-        const standardRouteModel = sidCollection.findRouteByIcao(sidId);
-        const exit = airport.getSIDExitPoint(sidId);
-        // TODO: perhaps this should use the `RouteModel`?
-        const route = `${airport.icao}.${sidId}.${exit}`;
-
-        if (_isNil(standardRouteModel)) {
-            return ['fail', 'SID name not understood'];
-        }
-
-        if (this.category !== FLIGHT_CATEGORY.DEPARTURE) {
-            return ['fail', 'unable to fly SID, we are an inbound'];
-        }
-
-        if (!this.rwy_dep) {
-            this.setDepartureRunway(airportController.airport_get().runway);
-        }
-
-        if (!standardRouteModel.hasFixName(this.rwy_dep)) {
-            return ['fail', `unable, the ${standardRouteModel.name} departure not valid from Runway ${this.rwy_dep}`];
-        }
-
-        // TODO: this is the wrong place for this `.toUpperCase()`
-        this.fms.followSID(route.toUpperCase());
-
-        const readback = {
-            log: `cleared to destination via the ${sidId} departure, then as filed`,
-            say: `cleared to destination via the ${standardRouteModel.name} departure, then as filed`
-        };
-
-        return ['ok', readback];
-    }
-
-    /**
-     * @for AircraftInstanceModel
-     * @method runSTAR
-     * @param data {array<string>} a string representation of the STAR, ex: `QUINN.BDEGA2.KSFO`
-     */
-    runSTAR(data) {
-        const routeModel = new RouteModel(data[0]);
-        const airport = window.airportController.airport_get();
-        const { name: starName } = airport.starCollection.findRouteByIcao(routeModel.procedure);
-
-        if (this.category !== FLIGHT_CATEGORY.ARRIVAL) {
-            return ['fail', 'unable to fly STAR, we are a departure!'];
-        }
-
-        // TODO: the data[0].length check might not be needed. this is covered via the CommandParser when
-        // this method runs as the result of a command.
-        if (data[0].length === 0 || !airport.starCollection.hasRoute(routeModel.procedure)) {
-            return ['fail', 'STAR name not understood'];
-        }
-
-        this.fms.followSTAR(routeModel.routeCode);
-
-        // TODO: casing may be an issue here.
-        const readback = {
-            log: `cleared to the ${airport.name} via the ${routeModel.procedure} arrival`,
-            say: `cleared to the ${airport.name} via the ${starName} arrival`
-        };
-
-        return ['ok', readback];
-    }
-
-    /**
-     * @for AircraftInstanceModel
-     * @method runMoveDataBlock
-     * @param data
-     */
-    runMoveDataBlock(dir) {
-        // TODO: what do all these numbers mean?
-        const positions = { 8: 360, 9: 45, 6: 90, 3: 135, 2: 180, 1: 225, 4: 270, 7: 315, 5: 'ctr' };
-
-        if (!_has(positions, dir[0])) {
-            return;
-        }
-
-        this.datablockDir = positions[dir[0]];
-    }
-
-    /**
-     * Adds a new Leg to fms with a user specified route
-     * Note: See notes on 'runReroute' for how to format input for this command
-     *
-     * @for AircraftInstanceModel
-     * @method runRoute
-     * @param data
-     */
-    runRoute(data) {
-         // capitalize everything
-        data = data[0].toUpperCase();
-        let worked = true;
-        const route = this.fms.formatRoute(data);
-
-        if (worked && route) {
-            // Add to fms
-            worked = this.fms.customRoute(route, false);
-        }
-
-        if (!route || !data || data.indexOf(' ') > -1) {
-            worked = false;
-        }
-
-        // Build the response
-        if (worked) {
-            const readback = {
-                log: `rerouting to :${this.fms.fp.route.join(' ')}`,
-                say: 'rerouting as requested'
-            };
-
-            return ['ok', readback];
-        }
-
-        const readback = {
-            log: `your route "${data}" is invalid!`,
-            say: 'that route is invalid!'
-        };
-
-        return ['fail', readback];
-    }
-
-    /**
-      * Removes all legs, and replaces them with the specified route
-      * Note: Input data needs to be provided with single dots connecting all
-      * procedurally-linked points (eg KSFO.OFFSH9.SXC or SGD.V87.MOVER), and
-      * all other points that will be simply a fix direct to another fix need
-      * to be connected with double-dots (eg HLI..SQS..BERRA..JAN..KJAN)
-      *
-      * @for AircraftInstanceModel
-      * @method runReroute
-      * @param data
-      */
-    runReroute(data) {
-        // TODO: capitalize everything?
-        data = data[0].toUpperCase();
-        let worked = true;
-        const route = this.fms.formatRoute(data);
-
-        if (worked && route) {
-            // Reset fms
-            worked = this.fms.customRoute(route, true);
-        }
-
-        // TODO: what exactly are we checking here?
-        if (!route || !data || data.indexOf(' ') > -1) {
-            worked = false;
-        }
-
-        // Build the response
-        if (worked) {
-            const readback = {
-                log: `rerouting to: ${this.fms.fp.route.join(' ')}`,
-                say: 'rerouting as requested'
-            };
-
-            return ['ok', readback];
-        }
-
-        const readback = {
-            log: `your route "${data}" is invalid!`,
-            say: 'that route is invalid!'
-        };
-
-        return ['fail', readback];
-    }
-
-    /**
-     * @for AircraftInstanceModel
-     * @method runTaxi
-     * @param data
-     */
-    runTaxi(data) {
-        // TODO: all this if logic should be simplified or abstracted
-        if (this.category !== FLIGHT_CATEGORY.DEPARTURE) {
-            return ['fail', 'inbound'];
-        }
-
-        if (this.mode === FLIGHT_MODES.TAXI) {
-            return ['fail', `already taxiing to ${radio_runway(this.rwy_dep)}`];
-        }
-
-        if (this.mode === FLIGHT_MODES.WAITING) {
-            return ['fail', 'already waiting'];
-        }
-
-        if (this.mode !== FLIGHT_MODES.APRON) {
-            return ['fail', 'wrong mode'];
-        }
-
-        // Set the runway to taxi to
-        if (data[0]) {
-            if (window.airportController.airport_get().getRunway(data[0].toUpperCase())) {
-                this.setDepartureRunway(data[0].toUpperCase());
-            } else {
-                return ['fail', `no runway ${data[0].toUpperCase()}`];
-            }
-        }
-
-        // Start the taxi
-        this.taxi_start = window.gameController.game_time();
-        const runway = window.airportController.airport_get().getRunway(this.rwy_dep);
-
-        runway.addQueue(this);
-        this.mode = FLIGHT_MODES.TAXI;
-
-        const readback = {
-            log: `taxi to runway ${runway.name}`,
-            say: `taxi to runway ${radio_runway(runway.name)}`
-        };
-
-        return ['ok', readback];
-    }
-
-    /**
-     * @for AircraftInstanceModel
-     * @method runTakeoff
-     * @param data
-     */
-    runTakeoff(data) {
-        // TODO: all this if logic should be simplified or abstracted
-        if (this.category !== 'departure') {
-            return ['fail', 'inbound'];
-        }
-
-        if (!this.isOnGround()) {
-            return ['fail', 'already airborne'];
-        }
-        if (this.mode === FLIGHT_MODES.APRON) {
-            return ['fail', 'unable, we\'re still in the parking area'];
-        }
-        if (this.mode === FLIGHT_MODES.TAXI) {
-            return ['fail', `taxi to runway ${radio_runway(this.rwy_dep)} not yet complete`];
-        }
-        if (this.mode === FLIGHT_MODES.TAKEOFF) {
-            // FIXME: this is showing immediately after a to clearance.
-            return ['fail', 'already taking off'];
-        }
-
-        if (this.fms.altitudeForCurrentWaypoint() <= 0) {
-            return ['fail', 'no altitude assigned'];
-        }
-
-        const runway = window.airportController.airport_get().getRunway(this.rwy_dep);
-
-        if (runway.removeQueue(this)) {
-            this.mode = FLIGHT_MODES.TAKEOFF;
-            this.scoreWind('taking off');
-            this.takeoffTime = window.gameController.game_time();
-
-            if (this.fms.currentWaypoint.speed == null) {
-                this.fms.setCurrent({ speed: this.model.speed.cruise });
-            }
-
-            const wind = window.airportController.airport_get().getWind();
-            const wind_dir = round(radiansToDegrees(wind.angle));
-            const readback = {
-                // TODO: the wind_dir calculation should be abstracted
-                log: `wind ${round(wind_dir / 10) * 10} ${round(wind.speed)}, runway ${this.rwy_dep} , cleared for takeoff`,
-                say: `wind ${radio_spellOut(round(wind_dir / 10) * 10)} at ${radio_spellOut(round(wind.speed))}, runway ${radio_runway(this.rwy_dep)}, cleared for takeoff`
-            };
-
-            return ['ok', readback];
-        }
-
-        const waiting = runway.inQueue(this);
-
-        return ['fail', `number ${waiting} behind ${runway.queue[waiting - 1].getRadioCallsign()}`, ''];
-    }
-
-    runLanding(data) {
-        const variant = data[0];
-        const runway = window.airportController.airport_get().getRunway(data[1]);
-
-        if (!runway) {
-            return ['fail', `there is no runway ${radio_runway(data[1])}`];
-        }
-
-        this.setArrivalRunway(data[1].toUpperCase());
-        // tell fms to follow ILS approach
-        this.fms.followApproach('ils', this.rwy_arr, variant);
-
-        const readback = {
-            log: `cleared ILS runway ${this.rwy_arr} approach`,
-            say: `cleared ILS runway ${radio_runway(this.rwy_arr)} approach`
-        };
-
-        return ['ok', readback];
-    }
-
-    /**
-     * @for AircraftInstanceModel
-     * @method runAbort
-     * @param data
-     */
-    runAbort(data) {
-        // TODO: these ifs on `mode` should be converted to a switch
-        if (this.mode === FLIGHT_MODES.TAXI) {
-            this.mode = FLIGHT_MODES.APRON;
-            this.taxi_start = 0;
-
-            console.log('aborted taxi to runway');
-
-            const isWarning = true;
-            window.uiController.ui_log(`${this.getCallsign()} aborted taxi to runway`, isWarning);
-
-            return ['ok', 'taxiing back to terminal'];
-        } else if (this.mode === FLIGHT_MODES.WAITING) {
-            return ['fail', 'unable to return to the terminal'];
-        } else if (this.mode === FLIGHT_MODES.LANDING) {
-            this.cancelLanding();
-
-            const readback = {
-                log: `go around, fly present heading, maintain ${this.fms.altitudeForCurrentWaypoint()}`,
-                say: `go around, fly present heading, maintain ${radio_altitude(this.fms.altitudeForCurrentWaypoint())}`
-            };
-
-            return ['ok', readback];
-        } else if (this.mode === FLIGHT_MODES.CRUISE && this.fms.currentWaypoint.navmode === WAYPOINT_NAV_MODE.RWY) {
-            this.cancelLanding();
-
-            const readback = {
-                log: `cancel approach clearance, fly present heading, maintain ${this.fms.altitudeForCurrentWaypoint()}`,
-                say: `cancel approach clearance, fly present heading, maintain ${radio_altitude(this.fms.altitudeForCurrentWaypoint())}`
-            };
-
-            return ['ok', readback];
-        } else if (this.mode === FLIGHT_MODES.CRUISE && this.fms.currentWaypoint.navmode === WAYPOINT_NAV_MODE.FIX) {
-            this.cancelFix();
-
-            if (this.category === FLIGHT_CATEGORY.ARRIVAL) {
-                return ['ok', 'fly present heading, vector to final approach course'];
-            } else if (this.category === 'departure') {
-                return ['ok', 'fly present heading, vector for entrail spacing'];
-            }
-        }
-
-        // modes 'apron', 'takeoff', ('cruise' for some navmodes)
-        return ['fail', 'unable to abort'];
-    }
-
-    // FIXME: is this in use?
-    /**
-     * @for AircraftInstanceModel
-     * @method runDebug
-     */
-    runDebug() {
-        window.aircraft = this;
-        return ['ok', { log: 'in the console, look at the variable &lsquo;aircraft&rsquo;', say: '' }];
-    }
-
-    // FIXME: is this in use?
-    /**
-     * @for AircraftInstanceModel
-     * @method runDelete
-     */
-    runDelete() {
-        window.aircraftController.aircraft_remove(this);
-    }
-
     // TODO: move to `fms.cancelFix()`
     /**
      * @for AircraftInstanceModel
@@ -2089,7 +1040,7 @@ export default class Aircraft {
                 const turning_radius = km(this.speed) / 3600 * turning_time; // dist covered in the turn, km
                 const dist_to_localizer = offset[0] / sin(angle_diff); // dist from the localizer intercept point, km
                 const turn_early_km = 1;    // start turn 1km early, to avoid overshoots from tailwind
-                const should_attempt_intercept = (0 < dist_to_localizer && dist_to_localizer <= turning_radius + turn_early_km);
+                const should_attempt_intercept = (dist_to_localizer > 0 && dist_to_localizer <= turning_radius + turn_early_km);
                 const in_the_window = abs(offset_angle) < degreesToRadians(1.5);  // if true, aircraft will move to localizer, regardless of assigned heading
 
                 if (should_attempt_intercept || in_the_window) {  // time to begin turn
